PurchaseTime should be long, otherwise it will overflow int

+++ InAppService/Xamarin.InAppBilling/Model/Purchase.cs
@@ -1,4 +1,3 @@
-
 namespace Xamarin.InAppBilling.Model
 {
 	/// <summary>
@@ -10,7 +9,7 @@ namespace Xamarin.InAppBilling.Model
 		public string OrderId { get; set; }
 		public string ProductId { get; set; }
 		public string DeveloperPayload { get; set; }
-		public int PurchaseTime { get; set; }
+		public long PurchaseTime { get; set; }
 		public int PurchaseState { get; set; }
 		public string PurchaseToken { get; set; }
